Add files via upload
A simolator which get to file - txt file of variables and xls file with phone numbers, and create an url for sending in order to simulate sending of warning.
the sumulator have two options: manualy and automatic. In manualy simulation, you can choose almost all parameters of the warning for sending to the system, while in the automatic situation, the simulator choose all the parameters randomly.
In order to protect the real project, some links/url are missing, such as full address of file.